requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Settings y configuración
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
cryptography>=42.0.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# AI/ML
scikit-learn==1.3.2
joblib==1.3.2

# External APIs
httpx==0.25.2
google-ads==22.1.0
requests==2.31.0

# Background tasks
celery==5.3.4
flower==2.0.1

# Validation
email-validator==2.1.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1